FUCK YEAH IT WORKS!

+++ src/fsm/room-fs.glsl
@@ -6,7 +6,7 @@ in vec3 gco; /* vertex shader space coordinates */
 in vec3 gno; /* vertex shader normal */
 
 layout (location = 0) out vec3 nofrag;
-layout (location = 1) out vec2 matfrag;
+layout (location = 1) out uvec2 matfrag;
 
 void main() {
   vec3 eye  = inverse(view)[3].xyz; /* camera position */
@@ -21,5 +21,5 @@ void main() {
   */
 
   nofrag = gno;
-  matfrag = vec2(1., 0.);
+  matfrag = uvec2(1, 0);
 }
